Attach JWT to WebSocket upgrade request
We'd like to set the Authorization header of the upgrade request
that the WebSocket constructor sends, but that is not permitted.
However, we can pass a second argument to the constructor, which
becomes the value of a `sec-websocket-protocol` header. So we pass
the JWT as a second argument to the constructor to give the server
our token.